+++ installation/pastell.sql
@@ -303,6 +303,18 @@ CREATE TABLE `type_dossier` (
 	`definition` json NOT NULL,
 	PRIMARY KEY (`id_t`)
 )  ENGINE=MyISAM  ;
+CREATE TABLE `type_dossier_action` (
+	`id_a` int(11) NOT NULL AUTO_INCREMENT,
+	`id_u` int(11) NOT NULL,
+	`id_t` int(11) NOT NULL,
+	`action` varchar(64) NOT NULL,
+	`date` datetime NOT NULL,
+	`empreinte_sha256` varchar(128) NOT NULL,
+	`message` mediumtext NOT NULL,
+	`export_json` json NOT NULL,
+	PRIMARY KEY (`id_a`),
+	KEY `id_t` (`id_t`)
+)  ENGINE=MyISAM  ;
 CREATE TABLE `utilisateur` (
 	`id_u` int(11) NOT NULL AUTO_INCREMENT,
 	`email` varchar(128) NOT NULL,